Apollo Demo Deployment changes
