+++ client/pages/api/apollo-client.ts
@@ -2,7 +2,7 @@ import { ApolloClient, InMemoryCache } from "@apollo/client";
 import { relayStylePagination } from "@apollo/client/utilities";
 
 const client = new ApolloClient({
-  uri: "http://localhost:5000/graphql",
+  uri: "https://next-apollo-server-demo.herokuapp.com/graphql",
   cache: new InMemoryCache({
     typePolicies: {
       Query: {
