+++ server/package.json
@@ -3,7 +3,8 @@
   "version": "1.0.0",
   "description": "",
   "scripts": {
-    "start": "nodemon ./src/server.ts"
+    "start": "ts-node ./src/server.ts",
+    "dev": "nodemon ./src/server.ts"
   },
   "keywords": [],
   "author": "",
