+++ server/src/mock/index.ts
@@ -1,18 +1,35 @@
+import {
+  addMocksToSchema,
+  relayStylePaginationMock,
+} from "@graphql-tools/mock";
 import casual from "casual";
-import { MockList } from "@graphql-tools/mock";
+import { buildSchema, GraphQLSchema } from "graphql";
+import typeDefs from "../schema";
 
-export const PersonDataMock = {
-  Query: () => ({
-    getPersons: (root: any, args: any) => {
-      console.log(args);
-      return new MockList(20);
+const schema = buildSchema(typeDefs);
+let mockedSchema: GraphQLSchema;
+mockedSchema = addMocksToSchema({
+  schema,
+  mocks: {
+    Query: {
+      items: () => ({
+        edges: [...new Array(2000)].map((_, index) => ({
+          node: {
+            name: casual.name,
+            age: casual.integer(1, 120),
+            address: casual.address,
+            email: casual.email,
+            phone: casual.phone,
+          },
+        })),
+      }),
+    },
+  },
+  resolvers: (store) => ({
+    Query: {
+      items: relayStylePaginationMock(store),
     },
   }),
-  Person: () => ({
-    name: casual.name,
-    age: () => casual.integer(0, 120),
-    address: casual.address,
-    email: casual.email,
-    phone: casual.phone,
-  }),
-};
+});
+
+export default mockedSchema;
