+++ server/src/server.ts
@@ -1,22 +1,27 @@
+import { ApolloServerPluginLandingPageDisabled } from "apollo-server-core";
 import { ApolloServer } from "apollo-server-express";
-import { ApolloServerPluginDrainHttpServer } from "apollo-server-core";
 import express from "express";
 import http from "http";
-import schemaWithMocks from "./schema";
-import { PersonSchema } from "./schema";
-import { PersonDataMock } from "./mock";
+import schemaWithMocks from "./mock";
+const PORT = process.env.PORT || "5000";
 
-async function listen(port: number) {
+async function listen(port: string) {
   const app = express();
   const httpServer = http.createServer(app);
   const corsOptions = {
-    origin: ["http://localhost:3000", "https://studio.apollographql.com"],
+    origin: [
+      "http://localhost:3000",
+      "https://nextjs-apollo-5szlld6zy-yogigoodman.vercel.app",
+      "https://studio.apollographql.com",
+    ],
   };
+
+  app.get("/", (req, res) => {
+    res.send(`Server Ready at ${port}`);
+  });
   const server = new ApolloServer({
     schema: schemaWithMocks,
-    // typeDefs: PersonSchema,
-    // mocks: PersonDataMock,
-    // plugins: [ApolloServerPluginDrainHttpServer({ httpServer })],
+    plugins: [ApolloServerPluginLandingPageDisabled()],
   });
   await server.start();
 
@@ -32,8 +37,8 @@ async function listen(port: number) {
 
 async function main() {
   try {
-    await listen(5000);
-    console.log("Server is ready at http://localhost:5000/graphql");
+    await listen(PORT);
+    console.log(`Server is ready at ${PORT}`);
   } catch (err) {
     console.error("Error starting the node server", err);
   }
